Move slice interfaces to their slices
Before they were collected in state.ts, but mainly used in the slice.
Keeping them in the slice makes it easier if the shape of a slice is
changed.
