+++ src/Device/DeviceSelector/DeviceList/BrokenDevice.tsx
@@ -8,8 +8,8 @@ import React from 'react';
 import { useDispatch } from 'react-redux';
 
 import PseudoButton from '../../../PseudoButton/PseudoButton';
-import { Device as DeviceProps } from '../../../state';
 import { showDialog } from '../../BrokenDeviceDialog/brokenDeviceDialogSlice';
+import { Device as DeviceProps } from '../../deviceSlice';
 import BasicDeviceInfo from '../BasicDeviceInfo';
 
 import './broken-device.scss';
