+++ src/state.ts
@@ -4,17 +4,17 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-import {
-    Device as NrfdlDevice,
-    SerialPort,
-} from '@nordicsemiconductor/nrf-device-lib-js';
-import { AnyAction, ThunkDispatch } from '@reduxjs/toolkit';
-import type { AutoDetectTypes } from '@serialport/bindings-cpp';
-import { SerialPortOpenOptions } from 'serialport';
-import { LogEntry } from 'winston';
+import type { AnyAction, ThunkDispatch } from '@reduxjs/toolkit';
 
 import type { DocumentationState } from './About/documentationSlice';
 import type { ShortcutState } from './About/shortcutSlice';
+import type { AppLayout } from './App/appLayout';
+import type { BrokenDeviceDialog } from './Device/BrokenDeviceDialog/brokenDeviceDialogSlice';
+import type { DeviceAutoSelectState } from './Device/deviceAutoSelectSlice';
+import type { DeviceSetupState } from './Device/deviceSetupSlice';
+import type { DeviceState } from './Device/deviceSlice';
+import type { ErrorDialog } from './ErrorDialog/errorDialogSlice';
+import type { Log } from './Log/logSlice';
 
 export type TDispatch = ThunkDispatch<RootState, null, AnyAction>;
 
@@ -32,92 +32,4 @@ export interface RootState {
     documentation: DocumentationState;
     brokenDeviceDialog: BrokenDeviceDialog;
     shortcuts: ShortcutState;
-}
-
-export interface AppLayout {
-    isSidePanelVisible: boolean;
-    isLogVisible: boolean;
-    currentPane: number;
-    paneNames: string[];
-}
-
-export interface ErrorMessage {
-    message: string;
-    detail?: string;
-}
-
-export interface ErrorDialog {
-    isVisible: boolean;
-    messages: ErrorMessage[];
-    errorResolutions?: ErrorResolutions;
-}
-
-export interface ErrorResolutions {
-    [key: string]: () => void;
-}
-
-export interface Log {
-    autoScroll: boolean;
-    logEntries: LogEntry[];
-    isLoggingVerbose: boolean;
-}
-
-export interface DeviceState {
-    devices: Map<string, Device>;
-    deviceInfo: Device | null;
-    selectedSerialNumber: string | null;
-    readbackProtection: 'unknown' | 'protected' | 'unprotected';
-}
-
-export interface DeviceSetupState {
-    visible: boolean;
-    onUserInput?: (canceled: boolean, choice?: number) => void;
-    message: string;
-    progressMessage?: string;
-    choices?: string[];
-    progress?: number;
-}
-
-export interface DeviceAutoSelectState {
-    autoReselect: boolean;
-    device?: Device;
-    disconnectionTime?: number;
-    waitForDevice?: WaitForDevice;
-    autoReconnectTimeout?: NodeJS.Timeout;
-    lastArrivedDeviceId?: number;
-    arrivedButWrongWhen?: boolean;
-}
-
-export interface WaitForDevice {
-    timeout: number;
-    when: 'always' | 'applicationMode' | 'dfuBootLoaderMode' | 'sameTraits';
-    once: boolean;
-    onSuccess?: (device: Device) => void;
-    onFail?: (reason?: string) => void;
-}
-
-export interface DeviceInfo {
-    name?: string | null;
-    cores?: number;
-    icon: React.ElementType;
-    website: {
-        productPagePath?: string;
-        buyOnlineParams?: string;
-    };
-}
-
-export interface Device extends NrfdlDevice {
-    serialNumber: string;
-    boardVersion?: string;
-    nickname?: string;
-    serialport?: SerialPort;
-    favorite?: boolean;
-    id: number;
-    persistedSerialPortOptions?: SerialPortOpenOptions<AutoDetectTypes>;
-}
-
-export interface BrokenDeviceDialog {
-    isVisible: boolean;
-    description: string;
-    url: string;
 }
